requirements.txt
numpy==1.19.5
tensorflow==2.5.0
Pillow==8.2.0
